requirements.txt
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
django-cors-headers==4.3.1
